Abstract basic functionality of Ui and Parser to seperate classes
