+++ src/main/java/Storage.java
@@ -1,0 +1,5 @@
+public class Storage {
+    public Storage() {
+
+    }
+}
