+++ src/main/java/TaskList.java
@@ -1,0 +1,6 @@
+public class TaskList {
+
+    public TaskList() {
+
+    }
+}
